+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Revision history for aeson-combinators
 
+## 0.1.1.0 -- 2023-12-29
+* lower bound for aeson-parser dependecy to fix build of documentation in hackage
+
+## 0.1.1.0 -- 2023-12-28
+* Aeson 2.2.x compatibility
+
+## 0.1.1.0 -- 2023-09-03
+* `fromDecoder` utility to unwrap a `Decoder`'s function
+* `maybeKey` combinator
+
 ## 0.1.0.1 -- 2022-05-03
 * fix benchmark compatibility with ghc 9.2.2